backend/requirements.txt
aiofiles==0.4.0
certifi==2019.11.28
chardet==3.0.4
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2020.2.21
httptools==0.1.1
httpx==0.9.3
hyperframe==5.2.0
idna==2.9
multidict==4.7.5
PyJWT==1.7.1
PyMySQL==0.9.3
python-dotenv==0.11.0
rfc3986==1.3.2
sanic==19.12.2
Sanic-Cors==0.10.0.post2
Sanic-Plugins-Framework==0.9.1
sniffio==1.1.0
ujson==1.35
uvloop==0.14.0
websockets==8.1
